+++ lib/peddler/apis/notifications_v1/event_filter.rb
@@ -24,6 +24,9 @@ module Peddler
 
         # @return [Array<OrderChangeTypeEnum>]
         attribute?(:order_change_types, Array, from: "orderChangeTypes")
+
+        # @return [TrackingIdentifier]
+        attribute?(:tracking_identifier, TrackingIdentifier, from: "trackingIdentifier")
       end
     end
   end