+++ lib/peddler/notifications/shipment_tracking_milestone_changed.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+# This file is generated. Do not edit.
+
+module Peddler
+  module Notifications
+    # Shipment Tracking Milestone Change Notification Schema.
+    #
+    # Notification sent when a shipment tracking milestone changes.
+    module ShipmentTrackingMilestoneChanged
+      class << self
+        def parse(hash)
+          Notification.parse(hash)
+        end
+      end
+    end
+  end
+end